deprecate things; move to use updated where possible
